+++ inc/misc.h
@@ -9,6 +9,6 @@
 #ifndef MISC_H_
 #define MISC_H_
 
-void print_value(int* value, int length);
+void print_value(uint8_t* value, int length);
 
 #endif /* MISC_H_ */
